batman-adv: Keep hard interface neighbor list sorted
The upcoming neighborhood hashing will compute a hash over the MAC
address of all neighbors on an interface, from the smallest to the
largest one.

This patch keeps the hard interface neighbor list in order to ease
the hash computation later.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>

---

Changes in v4:
* none

Changes in v3:
* add missing compat code for hlist_add_behind_rcu()
  (thanks Sven!)
* add linux/string.h in this instead of the next patch,
  memcmp in originator.h was introduced here already

Changes in v2:
* Moved sorted storing of hardif neighbors to this separate patch
* fix rcu bug: use hlist_add_{head,behind}_rcu() instead of their
  non-rcu variants (and rebase on top of Sven's maint patch for the
  original hlist_add_head() )
(thanks Sven!)