batman-adv: Introduce packet type independent TVLV handler API
Instead of having one TVLV registration function for OGMs and one for
unicast packets, this patch adds a new, generic tvlv handler
registratiorn function, which simply uses a packet type as parameter.

For now, this patch only migrates the multicast and gateway tvlv
handlers to this new API, as these two have been tested and verified
to work by the author.

The benefits of this new API:

* A more unified TVLV handling
* Easier to add TVLV capabilities to any new, upcoming packet type
* Does not rely on linearized skb data

Further ToDos (for later patches):

* Insert handler hooks for unicast_tvlv packets too and
  migrate BATADV_TVLV_{DAT,NC,TT,ROAM} to new API after
  further testing.
* Remove old TVLV handler API

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>

---

Changes in v4:
* remove unnecessary bug.h include (not needed anymore
  since the removal of the wrappers in v3)

Changes in v3:
* remove unnecessary skb_reset_network_header() call
  shortly before kfree_skb() batadv_iv_ogm_receive()
* remove tvlv (un)pack_ctx wrappers
(thanks Sven!)

Changes in v2:
* add includes for linux/{bug,printk}.h
* update forward declaration to "struct sk_buff;" in tvlv.h
* kerneldoc:
  - change a "@tvlv_type" to "@packet_type",
  - fixed batadv_tvlv_ogm_unpack_ctx() parameter kerneldoc
(thanks Sven!)