batman-adv: aggregation packet queueing and transmission
This patch implements the queueing of any broadcast packet and its
following tranmission in an aggregation packet.

A broadcast packet is usually queued for up to 100ms (plus-minus
some jitter) or about 50ms on average.

However if the aggregation worker cannot keep up with emptying its
queue, a 10 to 20 times faster aggregation packet transmission rate is
used.

If an aggregation queue still runs full then broadcast packets are handed
over to the driver directly without aggregation, leaving it up to the
driver to either transmit or drop it, for instance due to congestion.

Signed-off-by: Linus Lüssing <linus.luessing@c0d3.blue>