batman-adv: Allow to drop packets from all other bla gateway
The scenario from a998bf5dfbd7 ("batman-adv: add detection for complex
bridge loops") can currently not be handled by batman-adv and BLA. Instead
it relies on some extra software or administrator which will handle the
complex loop by adjusting the network setup.

One possible adjustment for this problem is to disallow traffic from other
gateways - even when they are not in the same physical backbone group.

Signed-off-by: Sven Eckelmann <sven@narfation.org>

Forwarded: not-needed
 It is currently only a test for some OpenMesh specific problems