batman-adv: make number of (re)broadcasts configurable per hardif
This adds two netlink configuration options to configure the number of
(re)broadcasts to send on a hard interface. One is for batman-adv broadcast
packets originating from this node and the second one is for forwarded
batman-adv broadcast packets.

This allows to reduce or increase the number of (re)broadcasts to send
on a wireless interface, for instance (default: 3, other interfaces 1).
Or allows to filter foreign or own batman-adv broadcast packets when set
to zero.

Also, with DAT DHCP snooping, the gateway DHCP feature, a low DHCPv4 lease
timeout (smaller than BATADV_DAT_ENTRY_TIMEOUT, currently 5 min.) and
group-aware multicast snooping set up then batman-adv broadcast flooding
is in theory not needed anymore for unicast or multicast (excluding
broadcast) communication with IPv4 or IPv6. Setting num_bcasts_own/other
to zero then reduces overhead and the risk of Denial-of-Service flooding
in a public network.

Signed-off-by: Linus Lüssing <ll@simonwunderlich.de>